AIX-4245 - Update name to aixplain SDK
