+++ aixplain_sdk/utils/config.py
@@ -1,12 +1,6 @@
-__title__ = "aixplain"
-__description__ = "aiXplain Pipelines adds AI functions to software."
-__url__ = "https://github.com/aixplain/pipelines/tree/main/docs"
-__version__ = "0.0.5"
-__author__ = "Thiago Castro Ferriera, Duraikrishna Selvaraju and Lucas Pavanelli"
-__author_email__ = "thiago.ferreira@aixplain.com, krishna.durai@aixplain.com, lucas.pavanelli@aixplain.com"
-__license__ = "http://www.apache.org/licenses/LICENSE-2.0"
-__copyright__ = """
-Copyright 2022 The aiXplain pipeline authors
+"""
+Copyright 2022 The aiXplain SDK authors
+
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -19,3 +13,12 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
+
+import os
+
+PIPELINES_RUN_URL = os.getenv("PIPELINES_RUN_URL", "https://platform-api.aixplain.com/assets/pipeline/execution/run")
+MODELS_RUN_URL = os.getenv("MODELS_RUN_URL", "https://models.aixplain.com/api/v1/execute")
+# GET THE API KEY FROM CMD
+PIPELINE_API_KEY = os.getenv("PIPELINE_API_KEY", "")
+MODEL_API_KEY = os.getenv("MODEL_API_KEY", "")
+LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO")
